requirements.txt
rich>=1.1.5,<2.0.0
appdirs>=1.4.4,<2.0.0
keyring>=21.2.1,<23.0.0
jira>=2.0.0,<3.0.0
pyinquirer>=1.0.3,<2.0.0
python-dateutil>=2.8.1,<3.0.0
pyyaml>=5.3.1,<6.0.0
